Test frontend

+++ app/actions/notifier.js
@@ -3,7 +3,7 @@ import {
   REMOVE_NOTIFICATION
 } from './_constants'
 
-const defaultNotification = {
+export const defaultNotification = {
   type: 'info',
   ttl: 3000
 }
@@ -12,16 +12,19 @@ export function notify(notification) {
   return (dispatch, getState) => {
     const idLast = getState().notifications.idLast + 1
     notification.id = idLast
+    notification.values = notification.values || {}
     if (!notification.hasOwnProperty('type')) {
       notification.type = defaultNotification.type
     }
     if (!notification.hasOwnProperty('ttl')) {
       notification.ttl = defaultNotification.ttl
     }
-    if(notification.ttl > 0) {
-      notification.timer = setTimeout(() =>
-        dispatch(removeNotify(idLast)), notification.ttl
-      )
+    if(process.env.NODE_ENV !== 'test') {
+      if(notification.ttl > 0) {
+        notification.timer = setTimeout(() =>
+          dispatch(removeNotify(idLast)), notification.ttl
+        )
+      }
     }
     dispatch({
       type: CREATE_NOTIFICATION,
@@ -38,7 +41,7 @@ export function removeNotify(id, force) {
       if(elem.id !== id) {
         return true
       }
-      if(force) {
+      if(force && elem.timer) {
         clearTimeout(elem.timer)
       }
     })
